+++ packages/gamut-labs/src/experimental/AppHeader/AppHeaderElements/AppHeaderLink/index.tsx
@@ -2,21 +2,16 @@ import { Anchor, Box } from '@codecademy/gamut';
 import styled from '@emotion/styled';
 import React from 'react';
 
-import { focusStyles } from '../SharedStyles';
-import { AppHeaderLinkItem } from '../types';
+import { focusStyles, hoverStyles } from '../SharedStyles';
+import { AppHeaderClickHandler, AppHeaderLinkItem } from '../types';
 
-const AppHeaderLinkButtonOuter = styled(Anchor)(({ theme }) => {
-  return `
-  color: ${theme.colors.navy};
-  text-decoration: none;
+const AppHeaderLinkButtonOuter = styled(Anchor)`
+  color: ${({ theme }) => theme.colors.navy};
   display: block;
-  &:hover {
-    color: ${theme.colors.hyper};
-    text-decoration: none;
-  }
+  text-decoration: none;
+  ${hoverStyles}
   ${focusStyles}
 `;
-});
 
 const AppHeaderLinkButtonInner = styled(Box)`
   white-space: nowrap;
@@ -24,7 +19,7 @@ const AppHeaderLinkButtonInner = styled(Box)`
 
 export type AppHeaderLinkProps = {
   item: AppHeaderLinkItem;
-  onClick: (event: React.MouseEvent) => {}; // TODO switch to HeaderClickHandler
+  onClick: AppHeaderClickHandler;
 };
 
 export const AppHeaderLink: React.FC<AppHeaderLinkProps> = ({
@@ -35,7 +30,8 @@ export const AppHeaderLink: React.FC<AppHeaderLinkProps> = ({
     <AppHeaderLinkButtonOuter
       data-testid={item.dataTestId}
       href={item.href}
-      onClick={(event: React.MouseEvent) => onClick(event)} // TODO pass item through
+      onClick={(event) => onClick(event, item)}
+      target={item.newTab ? 'blank' : ''}
       variant="interface"
     >
       <AppHeaderLinkButtonInner