+++ packages/gamut-labs/src/experimental/AppHeader/AppHeaderElements/AppHeaderLogo/index.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 
 import { Logo } from '../../../../brand/Logo';
 import { focusStyles } from '../SharedStyles';
-import { AppHeaderLogoItem } from '../types';
+import { AppHeaderClickHandler, AppHeaderLogoItem } from '../types';
 
 const LogoButton = styled(Anchor)`
   display: inline-block;
@@ -19,7 +19,7 @@ const StyledLogo = styled(Logo)`
 
 export type AppHeaderLogoProps = {
   item: AppHeaderLogoItem;
-  onClick: (event: React.MouseEvent) => {}; // eventually this will be of type HeaderClickHandler and take in a second parameter "item" (the react node)
+  onClick: AppHeaderClickHandler;
 };
 
 export const AppHeaderLogo: React.FC<AppHeaderLogoProps> = ({
@@ -29,7 +29,7 @@ export const AppHeaderLogo: React.FC<AppHeaderLogoProps> = ({
   return (
     <LogoButton
       data-testid={item.dataTestId}
-      onClick={(event: React.MouseEvent) => onClick(event)}
+      onClick={(event) => onClick(event, item)}
       href={item.href}
     >
       {<StyledLogo type={item.pro ? 'proMono' : 'default'} height={27} />}