+++ packages/gamut-labs/src/experimental/AppHeader/AppHeaderElements/types.tsx
@@ -1,9 +1,17 @@
+import { GamutIconProps } from '@codecademy/gamut-icons';
+import { ReactNode } from 'react';
+
+export type AppHeaderItem =
+  | AppHeaderLogoItem
+  | AppHeaderLinkItem
+  | AppHeaderTextButtonItem
+  | AppHeaderFillButtonItem
+  | AppHeaderDropdownItem
+  | AppHeaderRenderElementItem;
+
 type AppHeaderBaseItem = {
   dataTestId?: string;
   id: string;
-  image?: string;
-  icon?: string;
-  text?: string;
 };
 
 export type AppHeaderLogoItem = AppHeaderBaseItem & {
@@ -15,12 +23,40 @@ export type AppHeaderLogoItem = AppHeaderBaseItem & {
 
 export type AppHeaderLinkItem = AppHeaderBaseItem & {
   href: string;
+  icon?: React.ComponentType<GamutIconProps>;
+  newTab?: boolean;
+  text: string;
   trackingTarget: string;
   type: 'link';
 };
 
+export type AppHeaderTextButtonItem = AppHeaderBaseItem & {
+  href: string;
+  text: string;
+  trackingTarget: string;
+  type: 'text-button';
+};
+
+export type AppHeaderFillButtonItem = AppHeaderBaseItem & {
+  href: string;
+  text: string;
+  trackingTarget: string;
+  type: 'fill-button';
+};
+
 export type AppHeaderDropdownItem = AppHeaderBaseItem & {
   popover: AppHeaderLinkItem[];
+  text: string;
   trackingTarget: string;
   type: 'dropdown';
 };
+
+export type AppHeaderRenderElementItem = AppHeaderBaseItem & {
+  renderElement: () => ReactNode;
+  type: 'render-element';
+};
+
+export type AppHeaderClickHandler = (
+  event: React.MouseEvent,
+  item: AppHeaderItem
+) => void;