+++ packages/gamut/src/AccordionArea/index.tsx
@@ -1,7 +1,7 @@
+import styled from '@emotion/styled';
 import { motion } from 'framer-motion';
-import React, { useLayoutEffect, useState } from 'react';
-
-import styles from './styles.module.scss';
+import React, { useState } from 'react';
+import { useIsomorphicLayoutEffect } from 'react-use';
 
 export type AccordionAreaProps = {
   children: React.ReactNode;
@@ -22,8 +22,8 @@ export type AccordionAreaProps = {
 const transitionDuration = 0.2;
 
 const variants = {
-  expanded: { height: '100%' },
-  folded: { height: 0 },
+  expanded: { height: 'auto' },
+  folded: { height: '0' },
 };
 
 export const AccordionArea: React.FC<AccordionAreaProps> = ({
@@ -34,7 +34,7 @@ export const AccordionArea: React.FC<AccordionAreaProps> = ({
 }) => {
   const [delayExpanded, setDelayExpanded] = useState(expanded);
 
-  useLayoutEffect(() => {
+  useIsomorphicLayoutEffect(() => {
     const handle = setTimeout(
       () => setDelayExpanded(expanded),
       transitionDuration * 1000
@@ -46,16 +46,19 @@ export const AccordionArea: React.FC<AccordionAreaProps> = ({
   return (
     <div className={className}>
       {top}
-      <motion.div
+      <StyledAccordionBody
         aria-expanded={expanded}
-        className={styles.accordionBody}
-        initial={false}
+        initial={expanded ? 'expanded' : 'folded'}
         animate={expanded ? 'expanded' : 'folded'}
         variants={variants}
         transition={{ duration: transitionDuration, ease: 'easeInOut' }}
       >
         {(expanded || delayExpanded) && children}
-      </motion.div>
+      </StyledAccordionBody>
     </div>
   );
 };
+
+const StyledAccordionBody = styled(motion.div)`
+  overflow: hidden;
+`;